+++ packages/core/src/js/feedback/defaults.ts
@@ -1,4 +1,4 @@
-import type { FeedbackButtonProps, FeedbackWidgetProps } from './FeedbackWidget.types';
+import type { FeedbackButtonProps, FeedbackWidgetProps, ScreenshotButtonProps } from './FeedbackWidget.types';
 import { feedbackAlertDialog } from './utils';
 
 const FORM_TITLE = 'Report a Bug';
@@ -12,6 +12,7 @@ const IS_REQUIRED_LABEL = '(required)';
 const SUBMIT_BUTTON_LABEL = 'Send Bug Report';
 const CANCEL_BUTTON_LABEL = 'Cancel';
 const TRIGGER_LABEL = 'Report a Bug';
+const TRIGGER_SCREENSHOT_LABEL = 'Take a Screenshot';
 const ERROR_TITLE = 'Error';
 const FORM_ERROR = 'Please fill out all required fields.';
 const EMAIL_ERROR = 'Please enter a valid email address.';
@@ -84,5 +85,10 @@ export const defaultConfiguration: Partial<FeedbackWidgetProps> = {
 
 export const defaultButtonConfiguration: Partial<FeedbackButtonProps> = {
   triggerLabel: TRIGGER_LABEL,
+  triggerAriaLabel: '',
+};
+
+export const defaultScreenshotButtonConfiguration: Partial<ScreenshotButtonProps> = {
+  triggerLabel: TRIGGER_SCREENSHOT_LABEL,
   triggerAriaLabel: '',
 };