+++ samples/react-native/src/Screens/ErrorsScreen.tsx
@@ -32,6 +32,7 @@ const ErrorsScreen = (_props: Props) => {
   const [showBadCode, setShowBadCode] = React.useState(false);
   const [isFeedbackVisible, setFeedbackVisible] = React.useState(false);
   const [isFeedbackButtonVisible, setFeedbackButtonVisible] = React.useState(false);
+  const [isScreenshotButtonVisible, setScreenshotButtonVisible] = React.useState(false);
 
   const errorBoundaryFallback: FallbackRender = ({ eventId }) => (
     <Text>Error boundary caught with event id: {eventId}</Text>
@@ -235,6 +236,18 @@ const ErrorsScreen = (_props: Props) => {
             } else {
               Sentry.showFeedbackButton();
               setFeedbackButtonVisible(true);
+            }
+          }}
+        />
+                <Button
+          title="Show/Hide Screenshot Button"
+          onPress={() => {
+            if (isScreenshotButtonVisible) {
+              Sentry.hideScreenshotButton();
+              setScreenshotButtonVisible(false);
+            } else {
+              Sentry.showScreenshotButton();
+              setScreenshotButtonVisible(true);
             }
           }}
         />